Move file to legacy directory
Took 26 seconds
