+++ Dad-Joke-API/src/main/java/org/falcon/legacy/ApiCall.java
@@ -1,4 +1,4 @@
-package org.falcon;
+package org.falcon.legacy;
 
 import java.io.IOException;
 import java.net.URI;
