+++ Dad-Joke-API/src/main/java/org/falcon/legacy/FileLineRetriever.java
@@ -1,4 +1,4 @@
-package org.falcon;
+package org.falcon.legacy;
 
 import java.io.File;
 import java.io.FileNotFoundException;
