+++ Dad-Joke-API/src/main/java/org/falcon/legacy/Joke.java
@@ -1,4 +1,4 @@
-package org.falcon;
+package org.falcon.legacy;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
